Refactor MotionProps into vanilla package
