+++ packages/framer-motion/src/animation/animate/single-value.ts
@@ -3,14 +3,14 @@ import {
     motionValue as createMotionValue,
     isMotionValue,
     MotionValue,
+    UnresolvedValueKeyframe,
     ValueAnimationTransition,
 } from "motion-dom"
-import { GenericKeyframesTarget } from "../../types"
 import { animateMotionValue } from "../interfaces/motion-value"
 
 export function animateSingleValue<V extends string | number>(
     value: MotionValue<V> | V,
-    keyframes: V | GenericKeyframesTarget<V>,
+    keyframes: V | UnresolvedValueKeyframe<V>[],
     options?: ValueAnimationTransition
 ): AnimationPlaybackControlsWithThen {
     const motionValue = isMotionValue(value) ? value : createMotionValue(value)
