+++ packages/framer-motion/src/animation/interfaces/visual-element-target.ts
@@ -1,3 +1,4 @@
+import type { TargetAndTransition } from "motion-dom"
 import {
     AnimationPlaybackControlsWithThen,
     frame,
@@ -7,7 +8,6 @@ import {
 import type { AnimationTypeState } from "../../render/utils/animation-state"
 import { setTarget } from "../../render/utils/setters"
 import type { VisualElement } from "../../render/VisualElement"
-import type { TargetAndTransition } from "../../types"
 import { addValueToWillChange } from "../../value/use-will-change/add-will-change"
 import { getOptimisedAppearId } from "../optimized-appear/get-appear-id"
 import { animateMotionValue } from "./motion-value"
