+++ packages/framer-motion/src/animation/sequence/types.ts
@@ -8,10 +8,9 @@ import {
     UnresolvedValueKeyframe,
 } from "motion-dom"
 import { Easing } from "motion-utils"
-import { GenericKeyframesTarget } from "../../types"
 
 export type ObjectTarget<O> = {
-    [K in keyof O]?: O[K] | GenericKeyframesTarget<O[K]>
+    [K in keyof O]?: O[K] | UnresolvedValueKeyframe[]
 }
 
 export type SequenceTime =
