+++ packages/framer-motion/src/gestures/types.ts
@@ -1,253 +1,1 @@
-import { Point } from "motion-utils"
-import { EventInfo } from "../events/types"
-import { VariantLabels } from "../motion/types"
-import { TargetAndTransition } from "../types"
-import { PanInfo } from "./pan/PanSession"
-
 export type RemoveEvent = () => void
-
-/**
- * @public
- */
-export interface FocusHandlers {
-    /**
-     * Properties or variant label to animate to while the focus gesture is recognised.
-     *
-     * ```jsx
-     * <motion.input whileFocus={{ scale: 1.2 }} />
-     * ```
-     */
-    whileFocus?: VariantLabels | TargetAndTransition
-}
-
-/**
- * Passed in to tap event handlers like `onTap` the `TapInfo` object contains
- * information about the tap gesture such as it‘s location.
- *
- * ```jsx
- * function onTap(event, info) {
- *   console.log(info.point.x, info.point.y)
- * }
- *
- * <motion.div onTap={onTap} />
- * ```
- *
- * @public
- */
-export interface TapInfo {
-    /**
-     * Contains `x` and `y` values for the tap gesture relative to the
-     * device or page.
-     *
-     * ```jsx
-     * function onTapStart(event, info) {
-     *   console.log(info.point.x, info.point.y)
-     * }
-     *
-     * <motion.div onTapStart={onTapStart} />
-     * ```
-     *
-     * @public
-     */
-    point: Point
-}
-
-/**
- * @public
- */
-export interface TapHandlers {
-    /**
-     * Callback when the tap gesture successfully ends on this element.
-     *
-     * ```jsx
-     * function onTap(event, info) {
-     *   console.log(info.point.x, info.point.y)
-     * }
-     *
-     * <motion.div onTap={onTap} />
-     * ```
-     *
-     * @param event - The originating pointer event.
-     * @param info - An {@link TapInfo} object containing `x` and `y` values for the `point` relative to the device or page.
-     */
-    onTap?(event: MouseEvent | TouchEvent | PointerEvent, info: TapInfo): void
-
-    /**
-     * Callback when the tap gesture starts on this element.
-     *
-     * ```jsx
-     * function onTapStart(event, info) {
-     *   console.log(info.point.x, info.point.y)
-     * }
-     *
-     * <motion.div onTapStart={onTapStart} />
-     * ```
-     *
-     * @param event - The originating pointer event.
-     * @param info - An {@link TapInfo} object containing `x` and `y` values for the `point` relative to the device or page.
-     */
-    onTapStart?(
-        event: MouseEvent | TouchEvent | PointerEvent,
-        info: TapInfo
-    ): void
-
-    /**
-     * Callback when the tap gesture ends outside this element.
-     *
-     * ```jsx
-     * function onTapCancel(event, info) {
-     *   console.log(info.point.x, info.point.y)
-     * }
-     *
-     * <motion.div onTapCancel={onTapCancel} />
-     * ```
-     *
-     * @param event - The originating pointer event.
-     * @param info - An {@link TapInfo} object containing `x` and `y` values for the `point` relative to the device or page.
-     */
-    onTapCancel?(
-        event: MouseEvent | TouchEvent | PointerEvent,
-        info: TapInfo
-    ): void
-
-    /**
-     * Properties or variant label to animate to while the component is pressed.
-     *
-     * ```jsx
-     * <motion.div whileTap={{ scale: 0.8 }} />
-     * ```
-     */
-    whileTap?: VariantLabels | TargetAndTransition
-
-    /**
-     * If `true`, the tap gesture will attach its start listener to window.
-     *
-     * Note: This is not supported publically.
-     */
-    globalTapTarget?: boolean
-}
-export type PanHandler = (event: Event, info: PanInfo) => void
-
-/**
- * @public
- */
-export interface PanHandlers {
-    /**
-     * Callback function that fires when the pan gesture is recognised on this element.
-     *
-     * **Note:** For pan gestures to work correctly with touch input, the element needs
-     * touch scrolling to be disabled on either x/y or both axis with the
-     * [touch-action](https://developer.mozilla.org/en-US/docs/Web/CSS/touch-action) CSS rule.
-     *
-     * ```jsx
-     * function onPan(event, info) {
-     *   console.log(info.point.x, info.point.y)
-     * }
-     *
-     * <motion.div onPan={onPan} />
-     * ```
-     *
-     * @param event - The originating pointer event.
-     * @param info - A {@link PanInfo} object containing `x` and `y` values for:
-     *
-     *   - `point`: Relative to the device or page.
-     *   - `delta`: Distance moved since the last event.
-     *   - `offset`: Offset from the original pan event.
-     *   - `velocity`: Current velocity of the pointer.
-     */
-    onPan?(event: PointerEvent, info: PanInfo): void
-
-    /**
-     * Callback function that fires when the pan gesture begins on this element.
-     *
-     * ```jsx
-     * function onPanStart(event, info) {
-     *   console.log(info.point.x, info.point.y)
-     * }
-     *
-     * <motion.div onPanStart={onPanStart} />
-     * ```
-     *
-     * @param event - The originating pointer event.
-     * @param info - A {@link PanInfo} object containing `x`/`y` values for:
-     *
-     *   - `point`: Relative to the device or page.
-     *   - `delta`: Distance moved since the last event.
-     *   - `offset`: Offset from the original pan event.
-     *   - `velocity`: Current velocity of the pointer.
-     */
-    onPanStart?(event: PointerEvent, info: PanInfo): void
-
-    /**
-     * Callback function that fires when we begin detecting a pan gesture. This
-     * is analogous to `onMouseStart` or `onTouchStart`.
-     *
-     * ```jsx
-     * function onPanSessionStart(event, info) {
-     *   console.log(info.point.x, info.point.y)
-     * }
-     *
-     * <motion.div onPanSessionStart={onPanSessionStart} />
-     * ```
-     *
-     * @param event - The originating pointer event.
-     * @param info - An {@link EventInfo} object containing `x`/`y` values for:
-     *
-     *   - `point`: Relative to the device or page.
-     */
-
-    onPanSessionStart?(event: PointerEvent, info: EventInfo): void
-
-    /**
-     * Callback function that fires when the pan gesture ends on this element.
-     *
-     * ```jsx
-     * function onPanEnd(event, info) {
-     *   console.log(info.point.x, info.point.y)
-     * }
-     *
-     * <motion.div onPanEnd={onPanEnd} />
-     * ```
-     *
-     * @param event - The originating pointer event.
-     * @param info - A {@link PanInfo} object containing `x`/`y` values for:
-     *
-     *   - `point`: Relative to the device or page.
-     *   - `delta`: Distance moved since the last event.
-     *   - `offset`: Offset from the original pan event.
-     *   - `velocity`: Current velocity of the pointer.
-     */
-    onPanEnd?(event: PointerEvent, info: PanInfo): void
-}
-
-/**
- * @public
- */
-export interface HoverHandlers {
-    /**
-     * Properties or variant label to animate to while the hover gesture is recognised.
-     *
-     * ```jsx
-     * <motion.div whileHover={{ scale: 1.2 }} />
-     * ```
-     */
-    whileHover?: VariantLabels | TargetAndTransition
-
-    /**
-     * Callback function that fires when pointer starts hovering over the component.
-     *
-     * ```jsx
-     * <motion.div onHoverStart={() => console.log('Hover starts')} />
-     * ```
-     */
-    onHoverStart?(event: MouseEvent, info: EventInfo): void
-
-    /**
-     * Callback function that fires when pointer stops hovering over the component.
-     *
-     * ```jsx
-     * <motion.div onHoverEnd={() => console.log("Hover ends")} />
-     * ```
-     */
-    onHoverEnd?(event: MouseEvent, info: EventInfo): void
-}
