+++ packages/framer-motion/src/render/types.ts
@@ -1,6 +1,6 @@
+import type { AnimationDefinition } from "motion-dom"
 import { MotionValue } from "motion-dom"
 import type { Axis, Box } from "motion-utils"
-import { AnimationDefinition } from "../animation/types"
 import { ReducedMotionConfig } from "../context/MotionConfigContext"
 import type { PresenceContextProps } from "../context/PresenceContext"
 import { MotionProps } from "../motion/types"
@@ -71,66 +71,6 @@ export interface LayoutLifecycles {
      */
     onLayoutAnimationComplete?(): void
 }
-
-export interface AnimationLifecycles {
-    /**
-     * Callback with latest motion values, fired max once per frame.
-     *
-     * ```jsx
-     * function onUpdate(latest) {
-     *   console.log(latest.x, latest.opacity)
-     * }
-     *
-     * <motion.div animate={{ x: 100, opacity: 0 }} onUpdate={onUpdate} />
-     * ```
-     */
-    onUpdate?(latest: ResolvedValues): void
-
-    /**
-     * Callback when animation defined in `animate` begins.
-     *
-     * The provided callback will be called with the triggering animation definition.
-     * If this is a variant, it'll be the variant name, and if a target object
-     * then it'll be the target object.
-     *
-     * This way, it's possible to figure out which animation has started.
-     *
-     * ```jsx
-     * function onStart() {
-     *   console.log("Animation started")
-     * }
-     *
-     * <motion.div animate={{ x: 100 }} onAnimationStart={onStart} />
-     * ```
-     */
-    onAnimationStart?(definition: AnimationDefinition): void
-
-    /**
-     * Callback when animation defined in `animate` is complete.
-     *
-     * The provided callback will be called with the triggering animation definition.
-     * If this is a variant, it'll be the variant name, and if a target object
-     * then it'll be the target object.
-     *
-     * This way, it's possible to figure out which animation has completed.
-     *
-     * ```jsx
-     * function onComplete() {
-     *   console.log("Animation completed")
-     * }
-     *
-     * <motion.div
-     *   animate={{ x: 100 }}
-     *   onAnimationComplete={definition => {
-     *     console.log('Completed animating', definition)
-     *   }}
-     * />
-     * ```
-     */
-    onAnimationComplete?(definition: AnimationDefinition): void
-}
-
-export type EventProps = LayoutLifecycles & AnimationLifecycles
 
 export type CreateVisualElement<Instance> = (
     Component: string | React.ComponentType<React.PropsWithChildren<unknown>>,
