+++ packages/framer-motion/src/render/utils/animation-state.ts
@@ -1,10 +1,9 @@
+import type { AnimationDefinition, TargetAndTransition } from "motion-dom"
 import { VisualElementAnimationOptions } from "../../animation/interfaces/types"
 import { animateVisualElement } from "../../animation/interfaces/visual-element"
-import { AnimationDefinition } from "../../animation/types"
 import { isAnimationControls } from "../../animation/utils/is-animation-controls"
 import { isKeyframesTarget } from "../../animation/utils/is-keyframes-target"
 import { VariantLabels } from "../../motion/types"
-import { TargetAndTransition } from "../../types"
 import { shallowCompare } from "../../utils/shallow-compare"
 import type { VisualElement } from "../VisualElement"
 import { getVariantContext } from "./get-variant-context"
