+++ packages/framer-motion/src/render/utils/setters.ts
@@ -1,11 +1,10 @@
+import type {
+    AnimationDefinition,
+    UnresolvedValueKeyframe,
+    ValueKeyframesDefinition,
+} from "motion-dom"
 import { motionValue } from "motion-dom"
 import { isKeyframesTarget } from "../../animation/utils/is-keyframes-target"
-import {
-    SingleTarget,
-    TargetAndTransition,
-    TargetResolver,
-    ValueTarget,
-} from "../../types"
 import type { VisualElement } from "../VisualElement"
 import { resolveVariant } from "./resolve-dynamic-variants"
 
@@ -25,14 +24,16 @@ function setMotionValue(
     }
 }
 
-function resolveFinalValueInKeyframes(v: ValueTarget): SingleTarget {
+function resolveFinalValueInKeyframes(
+    v: ValueKeyframesDefinition
+): UnresolvedValueKeyframe {
     // TODO maybe throw if v.length - 1 is placeholder token?
     return isKeyframesTarget(v) ? v[v.length - 1] || 0 : v
 }
 
 export function setTarget(
     visualElement: VisualElement,
-    definition: string | TargetAndTransition | TargetResolver
+    definition: AnimationDefinition
 ) {
     const resolved = resolveVariant(visualElement, definition)
     let { transitionEnd = {}, transition = {}, ...target } = resolved || {}
