+++ packages/motion-dom/src/node/types.ts
@@ -1,0 +1,970 @@
+import type { BoundingBox, Box, Point } from "motion-utils"
+import { RefObject } from "react"
+import type {
+    DOMKeyframesDefinition,
+    InertiaOptions,
+    TransformProperties,
+    Transition,
+} from "../animation/types"
+import type { DragElastic } from "../gestures/drag/types"
+import { PanInfo } from "../gestures/pan/types"
+import type { MotionValue } from "../value"
+
+export interface EventInfo {
+    point: Point
+}
+
+/**
+ * A generic set of string/number values
+ */
+export interface ResolvedValues {
+    [key: string]: string | number
+}
+
+export type AnimationDefinition =
+    | VariantLabels
+    | TargetAndTransition
+    | TargetResolver
+
+/**
+ * An object that specifies values to animate to. Each value may be set either as
+ * a single value, or an array of values.
+ *
+ * It may also option contain these properties:
+ *
+ * - `transition`: Specifies transitions for all or individual values.
+ * - `transitionEnd`: Specifies values to set when the animation finishes.
+ *
+ * ```jsx
+ * const target = {
+ *   x: "0%",
+ *   opacity: 0,
+ *   transition: { duration: 1 },
+ *   transitionEnd: { display: "none" }
+ * }
+ * ```
+ *
+ * @public
+ */
+export type TargetAndTransition = DOMKeyframesDefinition & {
+    transition?: Transition
+    transitionEnd?: ResolvedValues
+}
+
+export type TargetResolver = (
+    custom: any,
+    current: ResolvedValues,
+    velocity: ResolvedValues
+) => TargetAndTransition | string
+
+/**
+ * Either a string, or array of strings, that reference variants defined via the `variants` prop.
+ * @public
+ */
+export type VariantLabels = string | string[]
+
+export type Variant = TargetAndTransition | TargetResolver
+
+export interface Variants {
+    [key: string]: Variant
+}
+
+export interface MotionNodeAnimationOptions {
+    /**
+     * Properties, variant label or array of variant labels to start in.
+     *
+     * Set to `false` to initialise with the values in `animate` (disabling the mount animation)
+     *
+     * ```jsx
+     * // As values
+     * <motion.div initial={{ opacity: 1 }} />
+     *
+     * // As variant
+     * <motion.div initial="visible" variants={variants} />
+     *
+     * // Multiple variants
+     * <motion.div initial={["visible", "active"]} variants={variants} />
+     *
+     * // As false (disable mount animation)
+     * <motion.div initial={false} animate={{ opacity: 0 }} />
+     * ```
+     */
+    initial?: TargetAndTransition | VariantLabels | boolean
+
+    /**
+     * Values to animate to, variant label(s), or `AnimationControls`.
+     *
+     * ```jsx
+     * // As values
+     * <motion.div animate={{ opacity: 1 }} />
+     *
+     * // As variant
+     * <motion.div animate="visible" variants={variants} />
+     *
+     * // Multiple variants
+     * <motion.div animate={["visible", "active"]} variants={variants} />
+     *
+     * // AnimationControls
+     * <motion.div animate={animation} />
+     * ```
+     */
+    animate?: TargetAndTransition | VariantLabels | boolean
+
+    /**
+     * A target to animate to when this component is removed from the tree.
+     *
+     * This component **must** be the first animatable child of an `AnimatePresence` to enable this exit animation.
+     *
+     * This limitation exists because React doesn't allow components to defer unmounting until after
+     * an animation is complete. Once this limitation is fixed, the `AnimatePresence` component will be unnecessary.
+     *
+     * ```jsx
+     * import { AnimatePresence, motion } from 'framer-motion'
+     *
+     * export const MyComponent = ({ isVisible }) => {
+     *   return (
+     *     <AnimatePresence>
+     *        {isVisible && (
+     *          <motion.div
+     *            initial={{ opacity: 0 }}
+     *            animate={{ opacity: 1 }}
+     *            exit={{ opacity: 0 }}
+     *          />
+     *        )}
+     *     </AnimatePresence>
+     *   )
+     * }
+     * ```
+     */
+    exit?: TargetAndTransition | VariantLabels
+
+    /**
+     * Variants allow you to define animation states and organise them by name. They allow
+     * you to control animations throughout a component tree by switching a single `animate` prop.
+     *
+     * Using `transition` options like `delayChildren` and `staggerChildren`, you can orchestrate
+     * when children animations play relative to their parent.
+
+     *
+     * After passing variants to one or more `motion` component's `variants` prop, these variants
+     * can be used in place of values on the `animate`, `initial`, `whileFocus`, `whileTap` and `whileHover` props.
+     *
+     * ```jsx
+     * const variants = {
+     *   active: {
+     *       backgroundColor: "#f00"
+     *   },
+     *   inactive: {
+     *     backgroundColor: "#fff",
+     *     transition: { duration: 2 }
+     *   }
+     * }
+     *
+     * <motion.div variants={variants} animate="active" />
+     * ```
+     */
+    variants?: Variants
+
+    /**
+     * Default transition. If no `transition` is defined in `animate`, it will use the transition defined here.
+     * ```jsx
+     * const spring = {
+     *   type: "spring",
+     *   damping: 10,
+     *   stiffness: 100
+     * }
+     *
+     * <motion.div transition={spring} animate={{ scale: 1.2 }} />
+     * ```
+     */
+    transition?: Transition
+}
+
+export interface MotionNodeEventOptions {
+    /**
+     * Callback with latest motion values, fired max once per frame.
+     *
+     * ```jsx
+     * function onUpdate(latest) {
+     *   console.log(latest.x, latest.opacity)
+     * }
+     *
+     * <motion.div animate={{ x: 100, opacity: 0 }} onUpdate={onUpdate} />
+     * ```
+     */
+    onUpdate?(latest: ResolvedValues): void
+
+    /**
+     * Callback when animation defined in `animate` begins.
+     *
+     * The provided callback will be called with the triggering animation definition.
+     * If this is a variant, it'll be the variant name, and if a target object
+     * then it'll be the target object.
+     *
+     * This way, it's possible to figure out which animation has started.
+     *
+     * ```jsx
+     * function onStart() {
+     *   console.log("Animation started")
+     * }
+     *
+     * <motion.div animate={{ x: 100 }} onAnimationStart={onStart} />
+     * ```
+     */
+    onAnimationStart?(definition: AnimationDefinition): void
+
+    /**
+     * Callback when animation defined in `animate` is complete.
+     *
+     * The provided callback will be called with the triggering animation definition.
+     * If this is a variant, it'll be the variant name, and if a target object
+     * then it'll be the target object.
+     *
+     * This way, it's possible to figure out which animation has completed.
+     *
+     * ```jsx
+     * function onComplete() {
+     *   console.log("Animation completed")
+     * }
+     *
+     * <motion.div
+     *   animate={{ x: 100 }}
+     *   onAnimationComplete={definition => {
+     *     console.log('Completed animating', definition)
+     *   }}
+     * />
+     * ```
+     */
+    onAnimationComplete?(definition: AnimationDefinition): void
+    onBeforeLayoutMeasure?(box: Box): void
+
+    onLayoutMeasure?(box: Box, prevBox: Box): void
+    onLayoutAnimationStart?(): void
+    onLayoutAnimationComplete?(): void
+}
+
+export interface MotionNodePanHandlers {
+    /**
+     * Callback function that fires when the pan gesture is recognised on this element.
+     *
+     * **Note:** For pan gestures to work correctly with touch input, the element needs
+     * touch scrolling to be disabled on either x/y or both axis with the
+     * [touch-action](https://developer.mozilla.org/en-US/docs/Web/CSS/touch-action) CSS rule.
+     *
+     * ```jsx
+     * function onPan(event, info) {
+     *   console.log(info.point.x, info.point.y)
+     * }
+     *
+     * <motion.div onPan={onPan} />
+     * ```
+     *
+     * @param event - The originating pointer event.
+     * @param info - A {@link PanInfo} object containing `x` and `y` values for:
+     *
+     *   - `point`: Relative to the device or page.
+     *   - `delta`: Distance moved since the last event.
+     *   - `offset`: Offset from the original pan event.
+     *   - `velocity`: Current velocity of the pointer.
+     */
+    onPan?(event: PointerEvent, info: PanInfo): void
+
+    /**
+     * Callback function that fires when the pan gesture begins on this element.
+     *
+     * ```jsx
+     * function onPanStart(event, info) {
+     *   console.log(info.point.x, info.point.y)
+     * }
+     *
+     * <motion.div onPanStart={onPanStart} />
+     * ```
+     *
+     * @param event - The originating pointer event.
+     * @param info - A {@link PanInfo} object containing `x`/`y` values for:
+     *
+     *   - `point`: Relative to the device or page.
+     *   - `delta`: Distance moved since the last event.
+     *   - `offset`: Offset from the original pan event.
+     *   - `velocity`: Current velocity of the pointer.
+     */
+    onPanStart?(event: PointerEvent, info: PanInfo): void
+
+    /**
+     * Callback function that fires when we begin detecting a pan gesture. This
+     * is analogous to `onMouseStart` or `onTouchStart`.
+     *
+     * ```jsx
+     * function onPanSessionStart(event, info) {
+     *   console.log(info.point.x, info.point.y)
+     * }
+     *
+     * <motion.div onPanSessionStart={onPanSessionStart} />
+     * ```
+     *
+     * @param event - The originating pointer event.
+     * @param info - An {@link EventInfo} object containing `x`/`y` values for:
+     *
+     *   - `point`: Relative to the device or page.
+     */
+
+    onPanSessionStart?(event: PointerEvent, info: EventInfo): void
+
+    /**
+     * Callback function that fires when the pan gesture ends on this element.
+     *
+     * ```jsx
+     * function onPanEnd(event, info) {
+     *   console.log(info.point.x, info.point.y)
+     * }
+     *
+     * <motion.div onPanEnd={onPanEnd} />
+     * ```
+     *
+     * @param event - The originating pointer event.
+     * @param info - A {@link PanInfo} object containing `x`/`y` values for:
+     *
+     *   - `point`: Relative to the device or page.
+     *   - `delta`: Distance moved since the last event.
+     *   - `offset`: Offset from the original pan event.
+     *   - `velocity`: Current velocity of the pointer.
+     */
+    onPanEnd?(event: PointerEvent, info: PanInfo): void
+}
+
+export interface MotionNodeHoverHandlers {
+    /**
+     * Properties or variant label to animate to while the hover gesture is recognised.
+     *
+     * ```jsx
+     * <motion.div whileHover={{ scale: 1.2 }} />
+     * ```
+     */
+    whileHover?: VariantLabels | TargetAndTransition
+
+    /**
+     * Callback function that fires when pointer starts hovering over the component.
+     *
+     * ```jsx
+     * <motion.div onHoverStart={() => console.log('Hover starts')} />
+     * ```
+     */
+    onHoverStart?(event: MouseEvent, info: EventInfo): void
+
+    /**
+     * Callback function that fires when pointer stops hovering over the component.
+     *
+     * ```jsx
+     * <motion.div onHoverEnd={() => console.log("Hover ends")} />
+     * ```
+     */
+    onHoverEnd?(event: MouseEvent, info: EventInfo): void
+}
+
+/**
+ * Passed in to tap event handlers like `onTap` the `TapInfo` object contains
+ * information about the tap gesture such as it‘s location.
+ *
+ * ```jsx
+ * function onTap(event, info) {
+ *   console.log(info.point.x, info.point.y)
+ * }
+ *
+ * <motion.div onTap={onTap} />
+ * ```
+ *
+ * @public
+ */
+export interface TapInfo {
+    /**
+     * Contains `x` and `y` values for the tap gesture relative to the
+     * device or page.
+     *
+     * ```jsx
+     * function onTapStart(event, info) {
+     *   console.log(info.point.x, info.point.y)
+     * }
+     *
+     * <motion.div onTapStart={onTapStart} />
+     * ```
+     *
+     * @public
+     */
+    point: Point
+}
+
+export interface MotionNodeTapHandlers {
+    /**
+     * Callback when the tap gesture successfully ends on this element.
+     *
+     * ```jsx
+     * function onTap(event, info) {
+     *   console.log(info.point.x, info.point.y)
+     * }
+     *
+     * <motion.div onTap={onTap} />
+     * ```
+     *
+     * @param event - The originating pointer event.
+     * @param info - An {@link TapInfo} object containing `x` and `y` values for the `point` relative to the device or page.
+     */
+    onTap?(event: MouseEvent | TouchEvent | PointerEvent, info: TapInfo): void
+
+    /**
+     * Callback when the tap gesture starts on this element.
+     *
+     * ```jsx
+     * function onTapStart(event, info) {
+     *   console.log(info.point.x, info.point.y)
+     * }
+     *
+     * <motion.div onTapStart={onTapStart} />
+     * ```
+     *
+     * @param event - The originating pointer event.
+     * @param info - An {@link TapInfo} object containing `x` and `y` values for the `point` relative to the device or page.
+     */
+    onTapStart?(
+        event: MouseEvent | TouchEvent | PointerEvent,
+        info: TapInfo
+    ): void
+
+    /**
+     * Callback when the tap gesture ends outside this element.
+     *
+     * ```jsx
+     * function onTapCancel(event, info) {
+     *   console.log(info.point.x, info.point.y)
+     * }
+     *
+     * <motion.div onTapCancel={onTapCancel} />
+     * ```
+     *
+     * @param event - The originating pointer event.
+     * @param info - An {@link TapInfo} object containing `x` and `y` values for the `point` relative to the device or page.
+     */
+    onTapCancel?(
+        event: MouseEvent | TouchEvent | PointerEvent,
+        info: TapInfo
+    ): void
+
+    /**
+     * Properties or variant label to animate to while the component is pressed.
+     *
+     * ```jsx
+     * <motion.div whileTap={{ scale: 0.8 }} />
+     * ```
+     */
+    whileTap?: VariantLabels | TargetAndTransition
+
+    /**
+     * If `true`, the tap gesture will attach its start listener to window.
+     *
+     * Note: This is not supported publically.
+     */
+    globalTapTarget?: boolean
+}
+
+export interface MotionNodeFocusHandlers {
+    /**
+     * Properties or variant label to animate to while the focus gesture is recognised.
+     *
+     * ```jsx
+     * <motion.input whileFocus={{ scale: 1.2 }} />
+     * ```
+     */
+    whileFocus?: VariantLabels | TargetAndTransition
+}
+
+/**
+ * TODO: Replace with types from inView()
+ */
+export type ViewportEventHandler = (
+    entry: IntersectionObserverEntry | null
+) => void
+
+export interface ViewportOptions {
+    root?: RefObject<Element | null>
+    once?: boolean
+    margin?: string
+    amount?: "some" | "all" | number
+}
+
+export interface MotionNodeViewportOptions {
+    whileInView?: VariantLabels | TargetAndTransition
+    onViewportEnter?: ViewportEventHandler
+    onViewportLeave?: ViewportEventHandler
+    viewport?: ViewportOptions
+}
+
+export interface MotionNodeDraggableOptions {
+    /**
+     * Enable dragging for this element. Set to `false` by default.
+     * Set `true` to drag in both directions.
+     * Set `"x"` or `"y"` to only drag in a specific direction.
+     *
+     * ```jsx
+     * <motion.div drag="x" />
+     * ```
+     */
+    drag?: boolean | "x" | "y"
+
+    /**
+     * Properties or variant label to animate to while the drag gesture is recognised.
+     *
+     * ```jsx
+     * <motion.div whileDrag={{ scale: 1.2 }} />
+     * ```
+     */
+    whileDrag?: VariantLabels | TargetAndTransition
+
+    /**
+     * If `true`, this will lock dragging to the initially-detected direction. Defaults to `false`.
+     *
+     * ```jsx
+     * <motion.div drag dragDirectionLock />
+     * ```
+     */
+    dragDirectionLock?: boolean
+
+    /**
+     * Allows drag gesture propagation to child components. Set to `false` by
+     * default.
+     *
+     * ```jsx
+     * <motion.div drag="x" dragPropagation />
+     * ```
+     */
+    dragPropagation?: boolean
+
+    /**
+     * Applies constraints on the permitted draggable area.
+     *
+     * It can accept an object of optional `top`, `left`, `right`, and `bottom` values, measured in pixels.
+     * This will define a distance from the named edge of the draggable component.
+     *
+     * Alternatively, it can accept a `ref` to another component created with React's `useRef` hook.
+     * This `ref` should be passed both to the draggable component's `dragConstraints` prop, and the `ref`
+     * of the component you want to use as constraints.
+     *
+     * ```jsx
+     * // In pixels
+     * <motion.div
+     *   drag="x"
+     *   dragConstraints={{ left: 0, right: 300 }}
+     * />
+     *
+     * // As a ref to another component
+     * const MyComponent = () => {
+     *   const constraintsRef = useRef(null)
+     *
+     *   return (
+     *      <motion.div ref={constraintsRef}>
+     *          <motion.div drag dragConstraints={constraintsRef} />
+     *      </motion.div>
+     *   )
+     * }
+     * ```
+     */
+    dragConstraints?: false | Partial<BoundingBox> | RefObject<Element | null>
+
+    /**
+     * The degree of movement allowed outside constraints. 0 = no movement, 1 =
+     * full movement.
+     *
+     * Set to `0.5` by default. Can also be set as `false` to disable movement.
+     *
+     * By passing an object of `top`/`right`/`bottom`/`left`, individual values can be set
+     * per constraint. Any missing values will be set to `0`.
+     *
+     * ```jsx
+     * <motion.div
+     *   drag
+     *   dragConstraints={{ left: 0, right: 300 }}
+     *   dragElastic={0.2}
+     * />
+     * ```
+     */
+    dragElastic?: DragElastic
+
+    /**
+     * Apply momentum from the pan gesture to the component when dragging
+     * finishes. Set to `true` by default.
+     *
+     * ```jsx
+     * <motion.div
+     *   drag
+     *   dragConstraints={{ left: 0, right: 300 }}
+     *   dragMomentum={false}
+     * />
+     * ```
+     */
+    dragMomentum?: boolean
+
+    /**
+     * Allows you to change dragging inertia parameters.
+     * When releasing a draggable Frame, an animation with type `inertia` starts. The animation is based on your dragging velocity. This property allows you to customize it.
+     * See {@link https://motion.dev/docs/react-motion-component#dragtransition | Inertia} for all properties you can use.
+     *
+     * ```jsx
+     * <motion.div
+     *   drag
+     *   dragTransition={{ bounceStiffness: 600, bounceDamping: 10 }}
+     * />
+     * ```
+     */
+    dragTransition?: InertiaOptions
+
+    /**
+     * Usually, dragging is initiated by pressing down on a component and moving it. For some
+     * use-cases, for instance clicking at an arbitrary point on a video scrubber, we
+     * might want to initiate dragging from a different component than the draggable one.
+     *
+     * By creating a `dragControls` using the `useDragControls` hook, we can pass this into
+     * the draggable component's `dragControls` prop. It exposes a `start` method
+     * that can start dragging from pointer events on other components.
+     *
+     * ```jsx
+     * const dragControls = useDragControls()
+     *
+     * function startDrag(event) {
+     *   dragControls.start(event, { snapToCursor: true })
+     * }
+     *
+     * return (
+     *   <>
+     *     <div onPointerDown={startDrag} />
+     *     <motion.div drag="x" dragControls={dragControls} />
+     *   </>
+     * )
+     * ```
+     */
+    dragControls?: any // TODO: Replace with DragControls when ported to motion-dom
+
+    /**
+     * If true, element will snap back to its origin when dragging ends.
+     *
+     * Enabling this is the equivalent of setting all `dragConstraints` axes to `0`
+     * with `dragElastic={1}`, but when used together `dragConstraints` can define
+     * a wider draggable area and `dragSnapToOrigin` will ensure the element
+     * animates back to its origin on release.
+     */
+    dragSnapToOrigin?: boolean
+
+    /**
+     * By default, if `drag` is defined on a component then an event listener will be attached
+     * to automatically initiate dragging when a user presses down on it.
+     *
+     * By setting `dragListener` to `false`, this event listener will not be created.
+     *
+     * ```jsx
+     * const dragControls = useDragControls()
+     *
+     * function startDrag(event) {
+     *   dragControls.start(event, { snapToCursor: true })
+     * }
+     *
+     * return (
+     *   <>
+     *     <div onPointerDown={startDrag} />
+     *     <motion.div
+     *       drag="x"
+     *       dragControls={dragControls}
+     *       dragListener={false}
+     *     />
+     *   </>
+     * )
+     * ```
+     */
+    dragListener?: boolean
+
+    /**
+     * If `dragConstraints` is set to a React ref, this callback will call with the measured drag constraints.
+     *
+     * @public
+     */
+    onMeasureDragConstraints?: (constraints: BoundingBox) => BoundingBox | void
+
+    /**
+     * Usually, dragging uses the layout project engine, and applies transforms to the underlying VisualElement.
+     * Passing MotionValues as _dragX and _dragY instead applies drag updates to these motion values.
+     * This allows you to manually control how updates from a drag gesture on an element is applied.
+     *
+     * @public
+     */
+    _dragX?: MotionValue<number>
+
+    /**
+     * Usually, dragging uses the layout project engine, and applies transforms to the underlying VisualElement.
+     * Passing MotionValues as _dragX and _dragY instead applies drag updates to these motion values.
+     * This allows you to manually control how updates from a drag gesture on an element is applied.
+     *
+     * @public
+     */
+    _dragY?: MotionValue<number>
+}
+
+export interface MotionNodeDragHandlers {
+    /**
+     * Callback function that fires when dragging starts.
+     *
+     * ```jsx
+     * <motion.div
+     *   drag
+     *   onDragStart={
+     *     (event, info) => console.log(info.point.x, info.point.y)
+     *   }
+     * />
+     * ```
+     *
+     * @public
+     */
+    onDragStart?(
+        event: MouseEvent | TouchEvent | PointerEvent,
+        info: PanInfo
+    ): void
+
+    /**
+     * Callback function that fires when dragging ends.
+     *
+     * ```jsx
+     * <motion.div
+     *   drag
+     *   onDragEnd={
+     *     (event, info) => console.log(info.point.x, info.point.y)
+     *   }
+     * />
+     * ```
+     *
+     * @public
+     */
+    onDragEnd?(
+        event: MouseEvent | TouchEvent | PointerEvent,
+        info: PanInfo
+    ): void
+
+    /**
+     * Callback function that fires when the component is dragged.
+     *
+     * ```jsx
+     * <motion.div
+     *   drag
+     *   onDrag={
+     *     (event, info) => console.log(info.point.x, info.point.y)
+     *   }
+     * />
+     * ```
+     *
+     * @public
+     */
+    onDrag?(event: MouseEvent | TouchEvent | PointerEvent, info: PanInfo): void
+
+    /**
+     * Callback function that fires a drag direction is determined.
+     *
+     * ```jsx
+     * <motion.div
+     *   drag
+     *   dragDirectionLock
+     *   onDirectionLock={axis => console.log(axis)}
+     * />
+     * ```
+     *
+     * @public
+     */
+    onDirectionLock?(axis: "x" | "y"): void
+
+    /**
+     * Callback function that fires when drag momentum/bounce transition finishes.
+     *
+     * ```jsx
+     * <motion.div
+     *   drag
+     *   onDragTransitionEnd={() => console.log('Drag transition complete')}
+     * />
+     * ```
+     *
+     * @public
+     */
+    onDragTransitionEnd?(): void
+}
+
+export interface MotionNodeLayoutOptions {
+    /**
+     * If `true`, this component will automatically animate to its new position when
+     * its layout changes.
+     *
+     * ```jsx
+     * <motion.div layout />
+     * ```
+     *
+     * This will perform a layout animation using performant transforms. Part of this technique
+     * involved animating an element's scale. This can introduce visual distortions on children,
+     * `boxShadow` and `borderRadius`.
+     *
+     * To correct distortion on immediate children, add `layout` to those too.
+     *
+     * `boxShadow` and `borderRadius` will automatically be corrected if they are already being
+     * animated on this component. Otherwise, set them directly via the `initial` prop.
+     *
+     * If `layout` is set to `"position"`, the size of the component will change instantly and
+     * only its position will animate.
+     *
+     * If `layout` is set to `"size"`, the position of the component will change instantly and
+     * only its size will animate.
+     *
+     * If `layout` is set to `"preserve-aspect"`, the component will animate size & position if
+     * the aspect ratio remains the same between renders, and just position if the ratio changes.
+     *
+     * @public
+     */
+    layout?: boolean | "position" | "size" | "preserve-aspect"
+
+    /**
+     * Enable shared layout transitions between different components with the same `layoutId`.
+     *
+     * When a component with a layoutId is removed from the React tree, and then
+     * added elsewhere, it will visually animate from the previous component's bounding box
+     * and its latest animated values.
+     *
+     * ```jsx
+     *   {items.map(item => (
+     *      <motion.li layout>
+     *         {item.name}
+     *         {item.isSelected && <motion.div layoutId="underline" />}
+     *      </motion.li>
+     *   ))}
+     * ```
+     *
+     * If the previous component remains in the tree it will crossfade with the new component.
+     *
+     * @public
+     */
+    layoutId?: string
+
+    /**
+     * A callback that will fire when a layout animation on this component starts.
+     *
+     * @public
+     */
+    onLayoutAnimationStart?(): void
+
+    /**
+     * A callback that will fire when a layout animation on this component completes.
+     *
+     * @public
+     */
+    onLayoutAnimationComplete?(): void
+
+    /**
+     * @public
+     */
+    layoutDependency?: any
+
+    /**
+     * Whether a projection node should measure its scroll when it or its descendants update their layout.
+     *
+     * @public
+     */
+    layoutScroll?: boolean
+
+    /**
+     * Whether an element should be considered a "layout root", where
+     * all children will be forced to resolve relatively to it.
+     * Currently used for `position: sticky` elements in Framer.
+     */
+    layoutRoot?: boolean
+
+    /**
+     * Attached to a portal root to ensure we attach the child to the document root and don't
+     * perform scale correction on it.
+     */
+    "data-framer-portal-id"?: string
+
+    /**
+     * By default, shared layout elements will crossfade. By setting this
+     * to `false`, this element will take its default opacity throughout the animation.
+     */
+    layoutCrossfade?: boolean
+}
+
+export type TransformTemplate = (
+    transform: TransformProperties,
+    generatedTransform: string
+) => string
+
+export interface MotionNodeAdvancedOptions {
+    /**
+     * Custom data to use to resolve dynamic variants differently for each animating component.
+     *
+     * ```jsx
+     * const variants = {
+     *   visible: (custom) => ({
+     *     opacity: 1,
+     *     transition: { delay: custom * 0.2 }
+     *   })
+     * }
+     *
+     * <motion.div custom={0} animate="visible" variants={variants} />
+     * <motion.div custom={1} animate="visible" variants={variants} />
+     * <motion.div custom={2} animate="visible" variants={variants} />
+     * ```
+     *
+     * @public
+     */
+    custom?: any
+
+    /**
+     * @public
+     * Set to `false` to prevent inheriting variant changes from its parent.
+     */
+    inherit?: boolean
+
+    /**
+     * @public
+     * Set to `false` to prevent throwing an error when a `motion` component is used within a `LazyMotion` set to strict.
+     */
+    ignoreStrict?: boolean
+
+    /**
+     * Provide a set of motion values to perform animations on.
+     */
+    values?: {
+        [key: string]: MotionValue<number> | MotionValue<string>
+    }
+
+    /**
+     * By default, Motion generates a `transform` property with a sensible transform order. `transformTemplate`
+     * can be used to create a different order, or to append/preprend the automatically generated `transform` property.
+     *
+     * ```jsx
+     * <motion.div
+     *   style={{ x: 0, rotate: 180 }}
+     *   transformTemplate={
+     *     ({ x, rotate }) => `rotate(${rotate}deg) translateX(${x}px)`
+     *   }
+     * />
+     * ```
+     *
+     * @param transform - The latest animated transform props.
+     * @param generatedTransform - The transform string as automatically generated by Motion
+     *
+     * @public
+     */
+    transformTemplate?: TransformTemplate
+
+    "data-framer-appear-id"?: string
+}
+
+export interface MotionNodeOptions
+    extends MotionNodeAnimationOptions,
+        MotionNodeEventOptions,
+        MotionNodePanHandlers,
+        MotionNodeTapHandlers,
+        MotionNodeHoverHandlers,
+        MotionNodeFocusHandlers,
+        MotionNodeViewportOptions,
+        MotionNodeDragHandlers,
+        MotionNodeDraggableOptions,
+        MotionNodeLayoutOptions,
+        MotionNodeAdvancedOptions {}
